+++ tests/test_high_level_api.py
@@ -10,3 +10,13 @@ def test_iterating_through_grib_file():
         for msg in grib:
             count += 1
     assert count == 1
+
+
+def test_iterating_through_grib_file_manual_close():
+    sample_path = os.path.join(codes_samples_path(), "GRIB2.tmpl")
+    with GribFile(sample_path) as grib:
+        count = 0
+        for msg in grib:
+            count += 1
+        msg.close()
+    assert count == 1